+++ cypress/e2e/spec-html.cy.js
@@ -16,7 +16,7 @@ const getIframeBody = () => {
 
 const makeNewFile = (filename = "new.html") => {
   cy.get("button[title='Project files']").click();
-  cy.get(".proj-new-component-button").click();
+  cy.get("span").contains("Add file").click();
   cy.get("div[class=modal-content__input]").find("input").type(filename);
   cy.get("div[class=modal-content__buttons]").contains("Add file").click();
 };