+++ src/containers/WebComponentLoader.jsx
@@ -1,6 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { disableTheming, setSenseHatAlwaysEnabled } from "../redux/EditorSlice";
+import {
+  disableTheming,
+  setSenseHatAlwaysEnabled,
+  triggerSave,
+} from "../redux/EditorSlice";
 import WebComponentProject from "../components/WebComponentProject/WebComponentProject";
 import { useTranslation } from "react-i18next";
 import { setInstructions } from "../redux/InstructionsSlice";
@@ -19,6 +23,7 @@ const WebComponentLoader = (props) => {
     code,
     senseHatAlwaysEnabled = false,
     instructions,
+    withProjectbar = false,
     withSidebar = false,
     sidebarOptions = [],
     theme,
@@ -41,6 +46,10 @@ const WebComponentLoader = (props) => {
   const themeDefault = window.matchMedia("(prefers-color-scheme:dark)").matches
     ? "dark"
     : "light";
+
+  useEffect(() => {
+    dispatch(triggerSave());
+  }, [dispatch]);
 
   useEffect(() => {
     if (theme) {
@@ -98,6 +107,7 @@ const WebComponentLoader = (props) => {
         }}
       >
         <WebComponentProject
+          withProjectbar={withProjectbar}
           withSidebar={withSidebar}
           sidebarOptions={sidebarOptions}
           hostStyles={hostStyles}